menambahkan detail invoice, kategori famili

+++ application/controllers/Dashboard.php
@@ -49,11 +49,17 @@ class Dashboard extends CI_Controller{
     }
 
     public function proses_pesanan(){
-        $this->cart->destroy();
-        $this->load->view('templates/header');
-        $this->load->view('templates/sidebar');
-        $this->load->view('proses_pesanan');
-        $this->load->view('templates/footer'); 
+        $is_processed = $this->Model_invoice->index();
+        if ($is_processed){
+            $this->cart->destroy();
+            $this->load->view('templates/header');
+            $this->load->view('templates/sidebar');
+            $this->load->view('proses_pesanan');
+            $this->load->view('templates/footer'); 
+        } else {
+            echo " Maaf, Pesanan Anda Gagal diproses!";
+        }
+      
     }
     public function detail($id){
         $ikan['data'] = $this->Model_ikan->find($id);
